requirements.txt
torch
diffusers
transformers
accelerate
streamlit
Pillow